requirements.txt
numpy==2.2.5
pandas==2.2.3
pyaerial==1.0.1
scikit_learn==1.6.1
setuptools==80.9.0
tabpfn==2.0.9
tabpfn_client==0.1.7
torch==2.7.0
tqdm==4.67.1
mlxtend==0.23.4
pyeclat==1.0.2